Start trimming comps
